+++ src/lib/filters/mask-filter.js
@@ -1,4 +1,6 @@
-export default class MaskFilter extends AbstractBaseFilter {
+import FoundryShim from "../../utils/foundry-shim.js";
+
+export default class MaskFilter extends FoundryShim.AbstractBaseFilter {
   /** @override */
   static fragmentShader = `\
         varying vec2 vTextureCoord;