+++ src/modules/sequencer-crosshair/CrosshairsDocument.js
@@ -1,6 +1,7 @@
 import CrosshairsPlaceable from "./CrosshairsPlaceable.js";
 import { get_object_canvas_data } from "../../lib/canvas-lib.js";
 import CONSTANTS from "../../constants.js";
+import FoundryShim from "../../utils/foundry-shim.js";
 
 /**
  *
@@ -37,6 +38,7 @@ export default class CrosshairsDocument extends MeasuredTemplateDocument {
 			snap: {
 				position: CONST.GRID_SNAPPING_MODES.CENTER,
 				size: CONST.GRID_SNAPPING_MODES.CENTER,
+				resolution: 1,
 				direction: 0
 			},
 			lockDrag: true,
@@ -106,13 +108,13 @@ export default class CrosshairsDocument extends MeasuredTemplateDocument {
 		const {t, distance, direction, angle, width} = this;
 		switch ( t ) {
 			case "circle":
-				return MeasuredTemplate.getCircleShape(distance);
+				return FoundryShim.MeasuredTemplate.getCircleShape(distance);
 			case "cone":
-				return MeasuredTemplate.getConeShape(distance, direction, angle);
+				return FoundryShim.MeasuredTemplate.getConeShape(distance, direction, angle);
 			case "rect":
-				return MeasuredTemplate.getRectShape(distance, direction);
+				return FoundryShim.MeasuredTemplate.getRectShape(distance, direction);
 			case "ray":
-				return MeasuredTemplate.getRayShape(distance, direction, width);
+				return FoundryShim.MeasuredTemplate.getRayShape(distance, direction, width);
 		}
 	}